+++ src/main/java/com/auth0/json/auth/TokenHolder.java
@@ -1,36 +1,43 @@
 package com.auth0.json.auth;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+
+import java.util.Date;
 
 /**
  * Class that contains the Tokens obtained after a call to the {@link com.auth0.client.auth.AuthAPI} methods.
  */
 @SuppressWarnings("unused")
-@JsonIgnoreProperties(ignoreUnknown = true)
-@JsonInclude(JsonInclude.Include.NON_NULL)
+@JsonDeserialize(using = TokenHolderDeserializer.class)
 public class TokenHolder {
 
-    @JsonProperty("access_token")
     private String accessToken;
-    @JsonProperty("id_token")
     private String idToken;
-    @JsonProperty("refresh_token")
     private String refreshToken;
-    @JsonProperty("token_type")
     private String tokenType;
-    @JsonProperty("expires_in")
     private long expiresIn;
-    @JsonProperty("scope")
     private String scope;
+    private Date expiresAt;
+
+    // We need to maintain a default constructor for backwards-compatibility
+    @SuppressWarnings("unused")
+    public TokenHolder() {}
+
+    public TokenHolder(String accessToken, String idToken, String refreshToken, String tokenType, long expiresIn, String scope, Date expiresAt) {
+        this.accessToken = accessToken;
+        this.idToken = idToken;
+        this.refreshToken = refreshToken;
+        this.tokenType = tokenType;
+        this.expiresIn = expiresIn;
+        this.scope = scope;
+        this.expiresAt = expiresAt;
+    }
 
     /**
      * Getter for the Auth0's access token.
      *
      * @return the access token or null if missing.
      */
-    @JsonProperty("access_token")
     public String getAccessToken() {
         return accessToken;
     }
@@ -40,7 +47,6 @@ public class TokenHolder {
      *
      * @return the id token or null if missing.
      */
-    @JsonProperty("id_token")
     public String getIdToken() {
         return idToken;
     }
@@ -50,7 +56,6 @@ public class TokenHolder {
      *
      * @return the refresh token or null if missing.
      */
-    @JsonProperty("refresh_token")
     public String getRefreshToken() {
         return refreshToken;
     }
@@ -60,19 +65,28 @@ public class TokenHolder {
      *
      * @return the token type or null if missing.
      */
-    @JsonProperty("token_type")
     public String getTokenType() {
         return tokenType;
     }
 
     /**
-     * Getter for the duration of this token in seconds since it was issued.
+     * Getter for the duration of the Access Token token in seconds from the time it was issued. If no Access Token was
+     * present the value will be zero.
      *
-     * @return the number of seconds in which this token will expire, since the time it was issued.
+     * @return the number of seconds in which the Access Token token will expire, from the time it was issued.
      */
-    @JsonProperty("expires_in")
     public long getExpiresIn() {
         return expiresIn;
+    }
+
+    /**
+     * Get the expiration date of the Access Token. This value is <strong>not</strong> part of the actual token response from the
+     * API, but rather is calculated and provided for convenience.
+     *
+     * @return the date of the Access Token's expiration. If no Access Token was present on the response, returns {@code null}
+     */
+    public Date getExpiresAt() {
+        return expiresAt;
     }
 
     /**
@@ -80,8 +94,8 @@ public class TokenHolder {
      *
      * @return a space-delimited string of the granted scopes of this token.
      */
-    @JsonProperty
     public String getScope() {
         return scope;
     }
+
 }